requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0.7
django-ckeditor==5.9.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-resized==0.3.11
furl==2.1.0
gunicorn==20.0.4
orderedmultidict==1.0.1
Pillow==7.1.2
psycopg2==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
whitenoise==5.1.0
